fix: Refactor data tab components and use Fireproof API properly
- Reorganized data components into the DataView directory
- Fixed DatabaseData component to use Fireproof API instead of direct IndexedDB access
- Moved DynamicTable and dynamicTableHelpers into DataView directory
- Updated imports and exports for better organization
- Added proper error handling and validation
