+++ app/components/ResultPreview/DataView/dynamicTableHelpers.ts
