+++ app/components/ResultPreview/DataView/index.ts
@@ -1,0 +1,5 @@
+// Export all data view components
+export { default as DatabaseData } from './DatabaseData';
+export { default as DatabaseListView } from './DatabaseListView';
+export { default as DynamicTable } from './DynamicTable';
+export { headersForDocs } from './dynamicTableHelpers';
